functions/ml_inference/requirements.txt
azure-functions
scikit-learn==1.7.2
xgboost==3.2.0
joblib
numpy
pandas
azure-storage-blob
azure-identity
